+++ src/api/axiosAuth.js
@@ -1,0 +1,12 @@
+import axios from 'axios';
+
+export const axiosAuth = () => {
+  const tokenObj = JSON.parse(localStorage.getItem('okta-token-storage'));
+  let tok = tokenObj.accessToken.accessToken;
+  return axios.create({
+    headers: {
+      Authorization: `Bearer ${tok}`,
+    },
+    baseURL: 'https://localhost:2019/',
+  });
+};
